r95: W1 session 词表重构对齐(alpha.4 27bf1039db refactor(session)! distinguish event seqs from log offsets)——①品牌双类型 SessionSeq/SessionLogOffset(type int64+NewSessionSeq/NewSessionLogOffset 构造器非负校验, 官方 SessionSeq must be a non-negative safe integer 语义, 与 persistence/revision.go Revision 品牌风格统一)+SessionSeqCursor/OptionalSessionSeq 别名; ②SessionHeader: SeedLength *int64 → IsSeeded bool+InheritedEventCount SessionLogOffset(官方 header 逻辑层迁移, index.ts:100 拒 seedLength 字段)+validateSessionHeader 一致性校验(unseeded cut 必须 0); ③磁盘 v0 兼容: jsonl toHeaderLine/headerFromLine 与 sqlite nullableSeedLength/rowToMeta 在 seedLength↔IsSeeded+InheritedEventCount 间双向翻译(官方 format.ts 'private version-0 physical header keeps numeric seedLength'——磁盘物理格式未变, SESSION_FORMAT_VERSION 官方仍 0 故不 bump, 红队建议④修正); ④Session 新方法 OwnEvents()/IsOwnSeq()/InheritedEventCount()(官方 ownEvents=snapshotEvents(inheritedEventCount)/isOwnSeq 语义)+FirstLiveSeq/Seq 返回类型品牌化为 SessionLogOffset(官方 get seq: SessionLogOffset); ⑤消费点迁移: agent/inbox.go(CreateAgentMeta SeedLength→IsSeeded+InheritedEventCount)+schedule(runtime/tools)+subagent(child-agent/continuation-manager/list-children/descriptor-seed)+sessiontelemetry+sessionquery+goal+projection+persistence(coordinator/storeadapter cursor 比较适配)+agentloop factory; ⑥全仓测试迁移(47+ 包, NewDetached/NewRestored 四参+header 字段适配, brand_test 新增品牌构造器/header 校验测试); 全量 go test ./... -count=1 默认环境 113+ 包 ok 0 FAIL

+++ goal/service.go
@@ -498,7 +498,7 @@ func (s *Service) cacheLocked(sess *session.Session) (*goalCache, error) {
 			return nil, err
 		}
 	}
-	cache := &goalCache{state: state, activation: ActivationDisarmed, observedSeq: sess.Seq()}
+	cache := &goalCache{state: state, activation: ActivationDisarmed, observedSeq: int64(sess.Seq())}
 	s.caches[sess] = cache
 	return cache, nil
 }
@@ -599,7 +599,7 @@ func (s *Service) commitSnapshotLocked(
 // state. The caller holds s.mu and emits the notification after releasing
 // it (listeners may reenter the service).
 func (s *Service) commitLocked(a *agent.Agent, cache *goalCache, change ChangeMeta, activation GoalActivation) error {
-	cache.pendingActivation = &pendingActivation{seq: a.Session.Seq(), activation: activation}
+	cache.pendingActivation = &pendingActivation{seq: int64(a.Session.Seq()), activation: activation}
 	defer func() { cache.pendingActivation = nil }()
 	if _, err := a.Session.Append(EventChange, change, nil); err != nil {
 		return err
